Refactor IsPatientPredicate
