+++ src/main/java/seedu/patientist/logic/commands/ListPatientsCommand.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 
 import seedu.patientist.commons.core.Messages;
 import seedu.patientist.model.Model;
-import seedu.patientist.model.person.IsPatientPredicate;
+import seedu.patientist.model.person.patient.IsPatientPredicate;
 
 /**
  * Lists all patients in the patientist book to the user.
@@ -12,9 +12,7 @@ import seedu.patientist.model.person.IsPatientPredicate;
 public class ListPatientsCommand extends Command {
     public static final String COMMAND_WORD = "lspat";
 
-    public static final String MESSAGE_USAGE = COMMAND_WORD + ": List all patients\n"
-           + "Parameters: KEYWORD [MORE_KEYWORDS]...\n"
-           + "Example: " + COMMAND_WORD + " alice bob charlie";
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Lists all patients";
     @Override
     public CommandResult execute(Model model) {
         requireNonNull(model);
